Polish "Migrate off deprecated Spring Data Cassandra API"
See gh-20662
